[iOS] Creating EGTests for Privacy Safe Browsing
This CL tests the normal workflow for interactions to and at the
Safe Browsing settings page. The normal workflow consists of navigating
to the settings menu > Privacy and Security > Safe Browsing and then
selecting a Safe Browsing option. Currently, there's an issue with the
accessibility activation point used in these TableViewItems so tests
that require tapping on the row or any of the "i" buttons have been
disabled until that issue is fixed.

Bug: 1307789, 1329733
Change-Id: Ie449da3df32fbe2fe6b51d13b6baf2bf9b7be593
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3687184
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1011525}
NOKEYCHECK=True
GitOrigin-RevId: 6c96b65225634c8d4ed9e231461352c6082f5b77