[UrlParamFilter][iOS] Add a TabHelper to track Cross OTR metrics.
"Cross OTR" refers to a change in navigation state from regular browsing
to Off-The-Record browsing. This CL is part of an experiment to remove certain parameters from a URL when opening a link into Off-The-Record browsing.

This TabHelper will be used to track metrics that occur after an
intervention occurs; in particular the HTTP response code and number of
refreshes that ensue.

Bug: 1324194
Change-Id: I138ee4b71fd8bb42932f07ee602a24d5ff6a8749
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3684490
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Matt Reichhoff <mreichhoff@chromium.org>
Commit-Queue: Kirubel Aklilu <kaklilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1011617}
NOKEYCHECK=True
GitOrigin-RevId: 255593a8d5bf9a9b41046d79084b74fb7c3d229a