[iOS] Cache selected overscroll action during async call
The linked bug has a case where
-overscrollActionsController:didTriggerAction: was called with NONE as
the action. This is guarded against, but because the actual call is
dispatched back to the main thread, it's possible that the
selectedAction changes between the time it is checked and used.

Instead, cache the selected action so it can't change during this time.

Bug: 1334024
Change-Id: I32609610c2cf2414ca83e1b0c25838e42aa3038f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3695623
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#1012469}
NOKEYCHECK=True
GitOrigin-RevId: 88b9126bc9dd1383154a3bad936d0304203f68c0