Return previous smart sorting ranking if no destinations are below fold.
The smart sorting algorithm, at a high-level, works by dividing the
destinations carousel into two groups: (A) visible "above-the-fold"
destinations and (B) non-visible "below-the-fold" destinations;
"below-the-fold" destinations are made visible to the user when they
scroll the carousel.

This CL adds functionality to short-circuit and not run the ranking
algorithm if there are no destinations below the fold (i.e. all
destinations are currently visible.)

Bug: 1334608,1322519
Change-Id: I9ba56c0f32b0dc61acdfbff1697f94843a1e40cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3696722
Commit-Queue: Benjamin Williams <bwwilliams@google.com>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#1012960}
NOKEYCHECK=True
GitOrigin-RevId: 5c866b176542df2330e7e29eeccd070e4ea673cc