[iOS] Always allow feed to be scrollable to minimum height
When feed is visible, checks the content size after model updates to
ensure that it meets the minimum height. If it doesn't, then it adds
content insets to the bottom of the collection view to make it
scrollable.

Change-Id: I49a9f8a439fac78abae9a6d301edd36da5c308a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3698117
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Adam Arcaro <adamta@google.com>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1016792}
NOKEYCHECK=True
GitOrigin-RevId: ad73034fa47b194ec7d8c3c1545b4c15e49f2079