[iOS] Improved accessibility functionality in ESB settings
Currently, the default double tap in VoiceOver for the rows in the
Enhanced Safe Browsing settings UI is to press the "i" button and the
custom action is used to select the row. This leads users to have a
non-intuitive workflow. To fix this, this CL switches these two
interactions so that the default double tap is used to select the row
and the custom action is used to press the "i" button. Additionally,
the accessibility hint has been updated to reflect these changes and
inform users that double tapping selects the option, not show more
information.

UI Video:
https://drive.google.com/file/d/1FzMw_5xHs4SrR_B2OjvoVhonoX-CsPxm/view?
usp=sharing&resourcekey=0-qVP0SGdnoTmCpRza1Okfow

Bug: 1327383
Change-Id: I30cd10191f19a1cfd0df8b417b02e7fd0b3803d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3714556
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1017710}
NOKEYCHECK=True
GitOrigin-RevId: 7eb26fcca23997a2c6a21cebaab60befe21230f0