[iOS] Fix PasswordsSettingsTestCase failures in iOS 16
The failures were caused by changes in the UIEditMenu. Multiple
elements were being found when searching for a grey matcher which were
the UIEditMenuCell and the UILabel. Changes were made to make the
matcher more specific and to target the UILabel instead of the
UIEditMenuCell. There's multiple elements, but only one that has
interaction enabled, so adding grey_userInteractionEnabled() fixes the problem.

Bug: 1331347
Change-Id: Ied523822f686818141d48fe9d20c5bf62f2d35f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3721879
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1018212}
NOKEYCHECK=True
GitOrigin-RevId: 0401e83a9bc58d82f87b02c7d473ecf72cd3b6d0