Always create JavaScriptContentWorld with a WKContentWorld
This ensures that the execution of JavaScript goes through WKFrameInfo
instead of `WebFrameImpl::EncryptPayload`.

The `__gCrWeb.message.routeMessage` API is not yet removed in this CL,
but a NOTREACHED() call is added at the point where this code is
called. Since it is no longer used, outdated tests which verify these
code paths are removed in this CL. The API itself will be removed in a
follow-up CL.

Bug: 1322482
Change-Id: I33eb36e3f0d26c902ca2541b1624817e5cca55ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3696560
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1018275}
NOKEYCHECK=True
GitOrigin-RevId: dd207fcee42f54aa219c5e1b61c7e7254018c5fb