[iOS] Exclude grayscale snapshots from iCloud backup
Grayscale snapshots are not useful to backup and restore. To save
space and network bandwidth, this CL prevents these images from
getting backed up. Since grayscale snapshots are stored in the same
directory as color snapshots, this exclusion is applied file-by-file
rather than to an entire directory, and is applied at the time that
a file is written. Existing snapshots for stale tabs are not affected
until they are re-written, to avoid repeatedly iterating over the
snapshot directory.

Bug: 1286752
Change-Id: Ibc16d5a731197e58b3dc2ec8ac9bcd99f9438a52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3730127
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1018444}
NOKEYCHECK=True
GitOrigin-RevId: 3b2d0e74315a016738c3970fab572d86331db35d