[Passwords] Cleanup after SupportForAddPasswordsInSettings launch
Remove the kSupportForAddPasswordsInSettings flag and use the guarded behavior by default on Desktop and iOS.

On iOS, the position of the edit button was also guarded by the flag.
The button is now put in the toolbar instead of the navigation bar.

Bug: 1226008
Change-Id: I114e367bfe461f4407981b74950bbaefc44157cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3710151
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Vidhan Jain <vidhanj@google.com>
Auto-Submit: Elizabeth Popova <lizapopova@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1018702}
NOKEYCHECK=True
GitOrigin-RevId: 83d6457aead309f3c49899b844900b7c092ceac3