[iOS] Re-enable test
Those bookmarks tests were failing because the cache position of the
bookmarks wasn't reset at the beginning of the test. Because of that,
if a test was failing after modifying the cache position of the
bookmarks, the subsequent tests are failing because they are assuming
a reset position.

This CL explicitly reset the cache position of the bookmarks at the
beginning of the test.

Bug: 1335808
Change-Id: I57aed1ec5ddd6fed00d29c4beffa22be54ca35b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3733115
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019020}
NOKEYCHECK=True
GitOrigin-RevId: a391aa5a1a49aeb47eb280e08a3441e85367f062