[ios] Try to decode HTTP headers as "utf-8" in addition to "latin1"
According to https://www.rfc-editor.org/rfc/rfc7230#section-3.2.4, the
HTTP headers have no specified encoding. NSHTTPURLResponse decode them
as "latin1" which results in garbage if the value is encoded in "utf-8".

Encoding value as "utf-8" is quite likely for values such as filename
property of "Content-Disposition". Interpreting those as "latin1" will
result in obviously incorrect filename (mojibake).

Add a workaround in CreateHeadersFromNSHTTPURLResponse() that tries to
decode the raw header content as "utf-8". The raw header value is not
accessible, but if it can be recreated by encoding the value returned
by NSHTTPURLResponse to "latin1" in an NSData.

Fixed: 1333351
Change-Id: I91f087cb8d4ef7eb18094e52549c482d236a04c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3735049
Reviewed-by: Ali Juma <ajuma@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019198}
NOKEYCHECK=True
GitOrigin-RevId: bd7921acd30a3de155187fe167fe4627fc6ce4fc