[iOS] Disable translate in full screen tests
Based on the screenshot of failures, there is a translate info bar
displayed. We supposed that when you scroll down, this info bar prevent
the top toolbar to be minimize. To fix the test, this CL disable
translate before the test.

Fixed: 1329265
Change-Id: I9cbe4532734159e8fa15968d2816cbfd63c86a1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3735225
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Louis Romero <lpromero@chromium.org>
Reviewed-by: Louis Romero <lpromero@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019520}
NOKEYCHECK=True
GitOrigin-RevId: 7d8c72cf001aa48a392bdaac83b0ccf9cf535aba