[iOS][UserPolicy]🏢 Managed account dialog strings in auth flow
We made a new set of title and subtitle strings for the managed
account dialog in the auth flow when the User Policy feature is enabled.

For now, the User Policy feature is conditionally enabled with the
--enable-user-policy-for-ios cmdline flag. This means that the new
strings are only used when the flag is explicitly flipped.

When User Policy is enabled it means that there will be an attempt to
fetch user policies but it doesn't necessarily mean that there will be
user policies that are active when showing the dialog (maybe
eventually or never if the admin hasn't set user policies).

Bug: 1341585
Change-Id: Ie3f734611488ecedb933be5ee1c6cf60f578e928
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3738952
Commit-Queue: Vincent Boisselle <vincb@google.com>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1020654}
NOKEYCHECK=True
GitOrigin-RevId: fd014eaf94470a90968353ffa26435837653de0f