[ios] Add a new trusted_vault API
Introduce a new class TrustedVaultClientBackend (exposing the same
API as ChromeTrustedVaultService) that is a KeyedService. Also add
a factory function to create a new instance.

This is not used yet as the internal repository need to be changed
to add an implementation of the provider API.

This is part of a multi CLs change:
 => https://crrev.com/c/3738185
 -  https://crrev.com/i/4854420
 -  https://crrev.com/c/3738606
 -  https://crrev.com/i/4854421
 -  https://crrev.com/c/3740648

Bug: 1201182
Change-Id: Ic71dd9beb3089a4b6e0360c316ca0931dd92dfc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3738185
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1020864}
NOKEYCHECK=True
GitOrigin-RevId: 016d2a5f54f9749c1c79fb342fd12abc87a881f7