[ios] Adjust for XCUIElement coordinate space changes in iOS 16.
On iOS 16, for some unknown reason, the XCUICoordinate space is scaled
by the device's scale factor when computing offsets relative to the
app or screen. This is likely an iOS bug, but work around it for now
to get tests passing again.

BUG=1342819

Change-Id: Ia024c56ef3c41d4d583a9550d6c6af69a756fca1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3751849
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1022124}
NOKEYCHECK=True
GitOrigin-RevId: 9aa97f4ddd5d49ec4b84436f604d9c983eba816c