[ios][feed] Add app agent
Previously, the DiscoverFeedService was created in the scene
controller. It should be an app agent, and this CL is simply
a refactor to move it from scene controller to an app agent.

The timing of the startup action has not changed.
Previous to this CL, the DiscoverFeedService was started in
InitStageFinal. This CL performs this action at the beginning
of (willTransitionTo) InitStageFinal. Effectively, these are
the same. The service has no interdependences with other
services that are run in InitStageFinal.

This CL is required for further improvements for the Feed.

Change-Id: Ia01aa3a855fa5856fd7e2cab677e8ce073592204
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3752061
Commit-Queue: edchin <edchin@google.com>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1023041}
NOKEYCHECK=True
GitOrigin-RevId: e125a5e930e8595b220ae94ee0f84fa4d5e206eb