[iOS] Add parameters to Incognito views and further refactoring
This CL makes revamped and non-revamped Incognito views more
customizable so they can be used in the incoming Incognito interstitial
for 3P intents. Specifically, it gives an option to show/hide the header
of the view i.e. top icon and main title.

It also fixes the non-revamped Incognito view to use the same strategy
as the revamped one for URL loading. View controllers should not have
access to model-layer objects such as `UrlLoadingBrowserAgent` either,
so a TODO comment is added with a link to the associated filed bug.

Bug: 1318750
Change-Id: Id18b02a62d1ca787f92ebe74d729847a67cb3128
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3698088
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Quentin Pubert <qpubert@google.com>
Cr-Commit-Position: refs/heads/main@{#1023741}
NOKEYCHECK=True
GitOrigin-RevId: 6d103bb8d206d025c3b85998ef634ae1e15514b0