[iOS] Add interstitial UI for 3P intents in Incognito
This CL adds the UI code for the Incognito interstitial promo style view. An Incognito interstitial coordinator managing a custom view
controller is added, as well as new strings and images.

The Incognito interstitial promo style view contains the Incognito NTP
component, which has been refactored in a different CL to make it
possible. It also has a navigation bar containing a Cancel button,
and whose background color is updated depending on the content offset
of the PromoStyleViewController's scroll view.

PromoStyleViewController is also updated with a new parameter, so the
top banner image can be constrained to fill the space horizontally and
look better in landscape mode, and to ensure the bounciness of the inner scroll view does not allow the user to see the void above the banner.

Bug: 1318750
Change-Id: I9f506de2a9c82b4e740259c8c5dcb676352a8bab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3686493
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Quentin Pubert <qpubert@google.com>
Reviewed-by: David Jean <djean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1023761}
NOKEYCHECK=True
GitOrigin-RevId: 51b79dea12e19db51ec7bcc938049b1793f1592c