[ios][UserPolicy]🏢 Show domains in chrome://management
The chrome://management page on iOS doesn't show any detail on the
domains that manage the account and the browser. We add new strings
to show these domains in the chrome://management page when they are
available.

There can be a domain for the browser policies and a domain for
user policies.

This is the implementation of the new chrome://management mocks
(http://shortn/_EK7WxuXF1K) for iOS.

Bug: 1325881, 1325877
Change-Id: I992af30a70419ccc613dc2fc50578d2b315d0fa2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3741937
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Vincent Boisselle <vincb@google.com>
Reviewed-by: Owen Min <zmin@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1023900}
NOKEYCHECK=True
GitOrigin-RevId: d331925db4f93c77d0dc78577868aa950078f0db