[FollowIPH] Don't show Follow IPH in incognito mode
Crash happens when a follow recommended page loaded in incognito mode
and we try to show an follow IPH, because the MetricsRecorder can't
be generated in an incognito browser.

To fix that now, we don't show the follow IPH in incognito mode.

WIP: checking more conditions before showing a follow IPH, and move
the checking logic to follow_iph_presenter_ so we won't need to
inject any pointer get from BrowserState in the future.

Bug: 1344217
Change-Id: Iec90d826415df353d996dfa455b9c17685139379
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3758369
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Tina Wang <tinazwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1023996}
NOKEYCHECK=True
GitOrigin-RevId: ca82576a2716429d5c311c673e36cb7bdf65e360