[ios] Fix FollowTabHelper
The base::CancelableTaskTracker passed to HistoryService needs to be
a member field as base::CancelableTaskTracker cancel the task in its
destructor which will happen when the method returns if it is used as
a local variable.

Having fixed that, the block passed to GetDailyVisitsToHost would
capture a pointer to `this` which is a C++ object and thus may come
to be dangling if the object is detroyed when the task result is
pending. The same issue exists for the GetFollowWebPageURLs call.

This is fixed by not using blocks but instead introducing helper
methods and binding them using WeakPtr<...> ensuring that the block
will not execute if the FollowTabHelper is deleted.

Use constants of type base::TimeDelta and initialize them with
helper function instead of encoding the unit in the variable name.

Convert a constant of type `std::string` to `const char[]` as using
global with a destructor is forbidden by the style guide.

Bug: none
Change-Id: I5acc05c6fb0bc8d995ca51da338841299363fc33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3766678
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Tina Wang <tinazwang@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025326}
NOKEYCHECK=True
GitOrigin-RevId: 481d808f3d5d4e6dd2a09690f59ea5b1140fb74d