Reland "[iOS] Re-enabling EG test for Enhanced Safe Browsing(ESB)"
This is a reland of commit 8f0c48ae7ac6a7507af30705ef9da18e3be9ab49.
This commit was originally reverted due to a bot failure where test
couldn't find the matcher because the screen on some simulators were
too small and required a scroll down before tapping the option. I
have made the necessary changes to make sure that the matcher is
found and tested that it works on the devices that were failing on the
bots and continued to work on devices that previously passed on test
runs.

Original change's description:
> [iOS] Re-enabling EG test for Enhanced Safe Browsing(ESB)
>
> The EG test was disabled due to the default activation point of
> TableViewInfoButtonItem's to be on the "i" button instead of the center
> of the row/cell. Since that has been resolved in the ESB menu, we can
> re-enable this test. Additionally, improvements were made to the
> EG test to make sure it worked successfully and to ensure that
> preference values were set properly in the EG test's set up and tear
> down.
>
> Bug: 1333625, 1307789, 1329733
> Change-Id: I005c1c33e52c7eb9a56966df817279886e74fd27
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3735159
> Reviewed-by: Gauthier Ambard <gambard@chromium.org>
> Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1020098}

Bug: 1333625, 1307789, 1329733, 1341363
Change-Id: I1ca447c2a9bf6347e876f20c4c5d5fb9ea2d6b1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3759156
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1025749}
NOKEYCHECK=True
GitOrigin-RevId: 96881f060a8bea906caed5a26063992f51b714f9