[CVC] Added presentation logic of CVC prompt
This patch adds logic to present the new CVC prompt. Most of the logic
is imported from the legacy prompt.

We will tackle the rearchitecture of the prompt based on Coordinators
and Commands after the new Component is launched, as it is the priority
to remove the dependency on deprecated Material Design Components first.

Bug for rearchitecture:
https://bugs.chromium.org/p/chromium/issues/detail?id=1344924

Based on CL:3760200

Fixes: 1344918
Change-Id: I33f322429007dccab36372c495899f262885da1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3765252
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Ernesto Izquierdo Clua <eic@google.com>
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1026462}
NOKEYCHECK=True
GitOrigin-RevId: 046223673be4acf86d8b51bd6cb9c0e0ec913b40