[FollowIPH] Check the IPH time interval for certain site (2/5)
To fulfill the follow IPH display condition:
Don't show IPH for a certain recommended site within a day,
we store a dictionary of NSURL -> Date pair in NSUserDefaults.

Each time when a website is recommended, we get the rss url of
this site and check if a website with the same rss url has
present follow IPH within a same day.

To do this, we add a new method to FollowProvider to return a
rss url when checking the recommended status. The existing
bool GetRecommendedStatus() need to be cleaned up after
internal change has been done.

The reason of using rss url is because websites that have
different urls can share a same rss url. Ex. cnn.com/news/news
and cnn.com/sports/another-news can have the same rss url, and
we don't want to display follow IPHs too often if the user keep
browsing pages starts with cnn.com/news/.

This is 2/5 a multi CLs change:
 -  cl/461945394
 => https://crrev.com/c/3774414
 -  https://crrev.com/i/4887361
 -  https://crrev.com/c/3774681 (cleaning up)
 -  cl/461991268 (cleaning up)

Bug: 1340154
Change-Id: Id0699b72b44e132c1dc82454d3dfa94a1fa77674
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3774414
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Tina Wang <tinazwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1026544}
NOKEYCHECK=True
GitOrigin-RevId: 53e2937cdf69105f0308655183e50bc19bf07506