adding strings for the "Share Chrome" feature
IDS_IOS_OVERFLOW_MENU_SHARE_CHROME is the text of the item in the
overflow menu. IDS_IOS_OVERFLOW_MENU_SHARE_CHROME_TITLE is the text
displayed in the preview of the app store URL.
IDS_IOS_OVERFLOW_MENU_SHARE_CHROME_DESC is the text that appears in the
payload of the shared message.

Change-Id: Ie1bbe403301ee1140e7061450a5e34690408499f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3785283
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1027846}
NOKEYCHECK=True
GitOrigin-RevId: ebdc9953f03f8b236ea8be48e9d677f7f98e3bf1