[iOS] Check if pages are the same for error pages
When checking if the current webpage is the same page as the failing
navigation, the WebClient should also be consulted.
This is preventing the page from creating a new navigation when the
user is reloading an offline page.

It used to be the case because the offline pages are changing the URL
of the page from url-failing.com to chrome-offline://url-failing.com
which means that when the check occurred, the URL were different,
creating a new navigation instead of changing the existing tab.

Fixed: 1198411
Change-Id: I151f008855b02424ccfd770e3c69a3236c88f758
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3748489
Commit-Queue: Ali Juma <ajuma@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1028273}
NOKEYCHECK=True
GitOrigin-RevId: bcff8bb82da1216da2064e1a29e19be8326913e2