[ios] DCHECK startup invariants
- UIApplication.applicationState is always UIApplicationStateBackground inside application:didFinishLaunchingWithOptions. So we can't rely
on that value.

- AppState applicationWillEnterForeground: will never be in InitStageStart because application:didFinishLaunchingWithOptions:
transitions past InitStageStart before returning to the runloop.

This CL adds DCHECKs to verify these invariants, but is meant to be
easily revertible if it causes an issue. A later CL should remove
these code paths.

This CL helps to reduce unused codepaths, in pursuit of properly
being able to cold start launch into the background.

Bug: 1346512
Change-Id: I8fa6d64ab7ba38bfc7420f142e3f89e4dca6b6cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3783457
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: edchin <edchin@google.com>
Cr-Commit-Position: refs/heads/main@{#1028527}
NOKEYCHECK=True
GitOrigin-RevId: 84d5f0cc8f8e51deb55b3af4c2d7d5cd235972a9