Ensure the OTR observer is alive for its client after shutdown
The OTR observer needs to be destroyed before browser state manager
since it observes the browser state and web states. The segmentation
service is a client of the OTR observer, but the service is destroyed
after the OTR observer is destroyed, when the browser state is cleared.
To fix this, add a TearDown() in OTR observer after which the observer
stops working and observing browser state, but stays alive for the
services that need to unregister observer clients.

BUG=1346593

Change-Id: I94edd95302dc5f7526df1b87d4351281db4e5d1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3786849
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Siddhartha S <ssid@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1029066}
NOKEYCHECK=True
GitOrigin-RevId: 278b6af50fe216266e6149810b74831bf7e80828