[ios] Activity services changes for "Share Chrome"
This CL makes 3 changes:
 * Adds ShareChrome to the ActivityScenario enum
 * Reenables support for additionalText when sharing a single URL
 * Allows giving an LPLinkMetadata when sharing a URL

History on additionalText: it appears that when support for sharing
multiple URLs was added, we stopped including additionalText in
the payload. For ShareChrome, we would like to include a very short
bit of text as part of the shared payload, so this change
reenables support for additionalText when sharing a single URL,
using the property that is still there.

Note on LPLinkMetadata: it seems that UIActivityViewController
is not able to produce a good preview of app store links. By using
LPLinkMetadata, we can provide the app icon and title to display
along with the URL.

PRD link: go/share-chrome-prd

Bug: 1340308
Change-Id: I1f4c1b80222e3b39419424dc5e2713221401515c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3783672
Commit-Queue: Scott Yoder <scottyoder@google.com>
Reviewed-by: Andrey Zaytsev <andzaytsev@google.com>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1029378}
NOKEYCHECK=True
GitOrigin-RevId: 3e87a7b31a09f8333fbf32ebff721523a73d95e1