[ios, Start] Update Return To Recent Tab when WebState's title changes
Since the NTP is being opened earlier in the startup process for Start,
it is more common for the page title to not be ready by the time Start
fetches it. Observing TitleWasSet() will help update the title when
it is ready.

Fixed: 1348937
Change-Id: I28f57556b7ccfd4928e53b4276eb0a7536e3d4ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3795125
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Auto-Submit: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1030139}
NOKEYCHECK=True
GitOrigin-RevId: 62bc0c34755a1403c82bf6bc12581af05458b7a6