Add AutofillSectionUponRedundantNameInfo to field trial config.
Depending on the configuration, bots pick up the feature from the field
trial config and enable it by default in browser tests like
FormStructureBrowserTest.DataDrivenHeuristics. As the expectations are
hardcoded in the output files for this test, the feature needs to be
enabled on all bots and is thus activated explicitly.

Bug: 1113970
Change-Id: Ia20dee18bf031d940de35d4de60998c2d8d24957
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3803191
Reviewed-by: Christoph Schwering <schwering@google.com>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1031063}
NOKEYCHECK=True
GitOrigin-RevId: a92622027e6116a4de6c42eff39cc7b8529d8d8d