[ios] Store "Share Chrome" feature usage in local prefs
This change adds two prefs: one to store the number of times the user
shares chrome, and another to store the time of the last share. The plan
is to use this data as a signal for other features, such as asking the
user to rate the app.

Bug: 1340308
Change-Id: I3652569ff3417c84a52c93d07429fb0aa54b3ff9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3806665
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/main@{#1031917}
NOKEYCHECK=True
GitOrigin-RevId: bf3db0d1a7dfb8079d32b26b13fc8785474f3e9a