[ios] Fixes persistent flake in CookiesTestCase.testClearIncognitoFromMain.
This test uses EG helpers to close all incognito tabs via model calls,
but as a result the tab grid is in an unexpected state (the main UI is
displaying normal tabs, while the tab grid is on the incognito
panel). A thumbstrip-specific codepath ends up incorrectly telling the
BVCContainerViewController to display nothing, resulting in an empty
black screen.

We can temporarily work around this failure by opening and closing the
tab grid -- this resets the current panel to show normal tabs, which
avoids the buggy codepaths.

Bug: 1350742
Change-Id: I248b8057440f2e98f0a55578ae25130f04f7df00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3813707
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1032300}
NOKEYCHECK=True
GitOrigin-RevId: 3dd57e7aadb8e0b5f0e1c3af85133b37bede165c