[iOS] Guard against reentrant tab inserts in the Tab Grid.
There are a few edge-case code paths where reentrant web state list
mutations are occurring (see the linked bug).

One common end point for these code paths is in the tab grid, where a
UI update causes the collection view to behave as if the new tab
button was pressed. The fix for that bug will be more complex, but
as a stopgap, this CL simply checks if the web state list is
being mutated when an insertion is triggered in the tab grid. If so,
it DCHECKs and safely no-ops in release builds. This should mitigate
many of this kind of crash while we fix the underlying bugs.

Bug: 1349357
Change-Id: I0256abea8f18a4f3d940f4ad77aaa5b00e5d2890
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3811157
Reviewed-by: David Jean <djean@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1032483}
NOKEYCHECK=True
GitOrigin-RevId: c0f37f1f1b8c54f01970bfbfaa244aede9658890