[iOS] Dismiss feed promo by tapping "X" on top right
Dismisses the feed top sync promo when the user taps X. The feed top
section always exists whether or not there is content in it. If it is
empty, it just has a height of 0. This prevents us from needing to
adjust constraints and adding/removing views dynamically.

Bug: 1331010
Change-Id: Ia8f833879fe1c3311a60a9e086992ad28fb0ee36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3804917
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Adam Arcaro <adamta@google.com>
Cr-Commit-Position: refs/heads/main@{#1032687}
NOKEYCHECK=True
GitOrigin-RevId: c927bc665f35481252984eab01ecae1d34d43e8b