[ios] Use new follow provider API
Introduce a FollowBrowserAgent that is used by the mediators to
communicate with the service. It register the callback that will
present the correct UI when the follow/unfollow request completes
and also record metrics.

As the FollowBrowserAgent is tied to a Browser, this fix the bug
where UI could be presented in the wrong window caused by the
old FollowProvider API.

This is part of a multi CLs change:
 -  https://crrev.com/c/3757671
 -  https://crrev.com/i/4874197
 => https://crrev.com/c/3760445
 -  https://crrev.com/i/4875937
 -  https://crrev.com/c/3758209

Bug: 1201182, 1342003
Change-Id: If9239db4587a20e0acd38b99700d9c18512b0b8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3760445
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Tina Wang <tinazwang@chromium.org>
Reviewed-by: Ian Wells <iwells@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1033020}
NOKEYCHECK=True
GitOrigin-RevId: 5552e45653cfee0fcb4d9fedc1db50e3183d221c