[iOS] Enable testsManagedAccountRemovedFromAnotherGoogleApp
This tests has the following scenario:
 * Sign-in and sync with a managed account
 * Simulate that this account disappear from SSOAuth
 * Verify that the user is signed out and their data has been removed

While removing the data, there is an activity overlay view. The
flakiness happens since EGTest doesn't wait for the overlay view
to be removed before to verify the state.

To fix the flakiness, now the test waits for the user to be signed out
and waits for the overlay view to be removed.

Fixed: 1348569
Change-Id: Ia4da868d18d8f288dc4e95bbb62e416fdaaa1a91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3810283
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1033444}
NOKEYCHECK=True
GitOrigin-RevId: 3b434acf76006ece2bff56241b42c3a452a7b125